Update assignment description for Feedback pull request workflow
Needed correct description of using the Feedback pull request
in the assignments.  Basically students should work on the `main`
branch, and push to main.  The Feedback pull request is set to
automatically track commits to main, and pull them to a `feedback`
branch.  Corrected assignment description to reflect the way
this works.