Update catch2 test runner extension
The extension for the catch2 tests seems to have just stopped
working.  Looks like there is a new api for VSCode test UI.
But for now changed to this fork of the original extension, which
still seems to be working.  Will need to investigate and
switch to the new test ui method and extensions in future.